fix: decouple bigquery types to resolve Netlify secrets scanner build error
