+++ src/components/MenuSection.tsx
@@ -1,6 +1,7 @@
 import Link from "next/link";
 import Image from "next/image";
 import { getSectionsData } from "@/lib/bigquery";
+// import { SectionData } from "@/lib/bigquery-types";
 import { getLocalizedSection } from "@/lib/i18n-db";
 import { getLocalizedUrl } from "@/lib/i18n-helpers";
